F_T30603 move MapBundle exceptions to core
