+++ demosplan/DemosPlanCoreBundle/Controller/Procedure/DemosPlanProcedureLayersAPIController.php
@@ -13,9 +13,9 @@ namespace demosplan\DemosPlanCoreBundle\Controller\Procedure;
 use DemosEurope\DemosplanAddon\Controller\APIController;
 use DemosEurope\DemosplanAddon\Response\APIResponse;
 use demosplan\DemosPlanCoreBundle\Annotation\DplanPermissions;
+use demosplan\DemosPlanCoreBundle\Exception\GisLayerCategoryTreeTooDeepException;
 use demosplan\DemosPlanCoreBundle\Exception\MessageBagException;
 use demosplan\DemosPlanCoreBundle\ResourceTypes\GisLayerCategoryResourceType;
-use demosplan\DemosPlanMapBundle\Exception\GisLayerCategoryTreeTooDeepException;
 use demosplan\DemosPlanMapBundle\Logic\MapHandler;
 use Exception;
 use Symfony\Component\HttpFoundation\JsonResponse;
