+++ demosplan/DemosPlanCoreBundle/Exception/GisLayerCategoryTreeTooDeepException.php
@@ -8,7 +8,7 @@
  * All rights reserved
  */
 
-namespace demosplan\DemosPlanMapBundle\Exception;
+namespace demosplan\DemosPlanCoreBundle\Exception;
 
 use Exception;
